Pagination

+++ core/repositories/repository.go
@@ -8,7 +8,7 @@ type Repository[T models.Model] interface {
 	All() []T
 	Count() int64
 	Find(id int64) *T
-	Limit(limit int, offset int) []T
+	Limit(limit int64, offset int64) []T
 	List(query string, args ...any) []T
 
 	Insert(entity any)
